New Debian revision 1.0-2.
This revision removes the ll2utm and utm2ll manpages from the source
package and no longer installs the binaries for ll2utm and utm2ll.

This is in response to piuparts complaints that these binaries conflict
with binaries in the drawmap package, and these binaries are not
critical to the use of direwolf. See bug #768223.